shift openInJosm modal handler content to api.js
